move api_not_found to ctj_api/urls.py
